Restarting Update
I am going to work on this and change the API to LWJGL instead of
JavaFX. Also if you know how to use Ant or Gradle or Maven, please tell
me because I really need to learn how to do it and for some reason I
just don't get it. I am not done, but this is probably the only thing
that I'll work on over winter break cause I got time to yeah, stay at
home in the cold and cry from winter depression. I finally finished the
second half of the Java courses, so I think I'm better off now with
abstract data structures now.